02apirunner/requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
cffi==1.15.1
chardet==3.0.4
cryptography==41.0.7
idna==2.10
importlib-metadata==6.7.0
iniconfig==2.0.0
jsonpath==0.82
packaging==23.2
paramiko==2.11.0
pluggy==0.13.1
py==1.11.0
pycparser==2.21
PyMySQL==0.10.1
PyNaCl==1.5.0
pytest==6.2.1
PyYAML==6.0.1
pyyaml-include==1.3.2
requests==2.24.0
six==1.16.0
soupsieve==2.4.1
toml==0.10.2
typing_extensions==4.7.1
urllib3==1.25.11
zipp==3.15.0
